:tada: feat: create ingredient unit can be double (#37)
* :tada: feat: create ingredient unit can be double

* :wrench: fix: เอาแก้ unit ออกก่อน

* :wrench: fix: ทำให้ update ingredient used ก็ต่อเมื่อ COOKING

+++ restaurant/src/main/java/ku/cs/restaurant/entity/Recipe.java
@@ -10,7 +10,7 @@ public class Recipe {
     private RecipeKey id;
 
     @Column(name = "qty")
-    private int qty;
+    private double qty;
 
     @ManyToOne
     @MapsId("ingredientId")
